models/userにcontrollerの処理を追記（models/postにあったgetUserPostをuserへ移動）（develop）
